functions/requirements.txt
firebase_functions~=0.1.0
openai
google-cloud-secret-manager
firebase_admin